requirements.txt
diffusers==0.3.0
einops==0.4.1
fire==0.4.0
numpy==1.22.3
Pillow==9.3.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.63.1
transformers==4.22.1
